+++ crates/matrix-sdk-crypto/src/types/events/utd_cause.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use matrix_sdk_common::deserialized_responses::{
-    UnableToDecryptInfo, UnableToDecryptReason, VerificationLevel,
+    UnableToDecryptInfo, UnableToDecryptReason, VerificationLevel, WithheldCode,
 };
 use ruma::{events::AnySyncTimelineEvent, serde::Raw, MilliSecondsSinceUnixEpoch};
 use serde::Deserialize;
@@ -57,6 +57,19 @@ pub enum UtdCause {
     /// be confused with pre-join or pre-invite messages (see
     /// [`UtdCause::SentBeforeWeJoined`] for that).
     HistoricalMessage = 5,
+
+    /// The keys for this event are intentionally withheld.
+    ///
+    /// The sender has refused to share the key because our device does not meet
+    /// the sender's security requirements.
+    WithheldForUnverifiedOrInsecureDevice = 6,
+
+    /// The keys for this event are missing, likely because the sender was
+    /// unable to share them (e.g., failure to establish an Olm 1:1
+    /// channel). Alternatively, the sender may have deliberately excluded
+    /// this device by cherry-picking and blocking it, in which case, no action
+    /// can be taken on our side.
+    WithheldBySender = 7,
 }
 
 /// MSC4115 membership info in the unsigned area.
@@ -97,8 +110,18 @@ impl UtdCause {
         unable_to_decrypt_info: &UnableToDecryptInfo,
     ) -> Self {
         // TODO: in future, use more information to give a richer answer. E.g.
-        match unable_to_decrypt_info.reason {
-            UnableToDecryptReason::MissingMegolmSession
+        match &unable_to_decrypt_info.reason {
+            UnableToDecryptReason::MissingMegolmSession { withheld_code: Some(reason) } => {
+                match reason {
+                    WithheldCode::Unverified => UtdCause::WithheldForUnverifiedOrInsecureDevice,
+                    WithheldCode::Blacklisted
+                    | WithheldCode::Unauthorised
+                    | WithheldCode::Unavailable
+                    | WithheldCode::NoOlm
+                    | WithheldCode::_Custom(_) => UtdCause::WithheldBySender,
+                }
+            }
+            UnableToDecryptReason::MissingMegolmSession { withheld_code: None }
             | UnableToDecryptReason::UnknownMegolmMessageIndex => {
                 // Look in the unsigned area for a `membership` field.
                 if let Some(unsigned) =
@@ -424,7 +447,7 @@ mod tests {
     fn missing_megolm_session() -> UnableToDecryptInfo {
         UnableToDecryptInfo {
             session_id: None,
-            reason: UnableToDecryptReason::MissingMegolmSession,
+            reason: UnableToDecryptReason::MissingMegolmSession { withheld_code: None },
         }
     }