+++ crates/matrix-sdk-crypto/src/types/mod.rs
@@ -34,6 +34,7 @@ use std::{
 };
 
 use as_variant::as_variant;
+use matrix_sdk_common::deserialized_responses::PrivOwnedStr;
 use ruma::{
     serde::StringEnum, DeviceKeyAlgorithm, DeviceKeyId, OwnedDeviceKeyId, OwnedUserId, UserId,
 };
@@ -422,20 +423,6 @@ impl Algorithm for OwnedDeviceKeyId {
 impl Algorithm for DeviceKeyAlgorithm {
     fn algorithm(&self) -> DeviceKeyAlgorithm {
         self.to_owned()
-    }
-}
-
-// Wrapper around `Box<str>` that cannot be used in a meaningful way outside of
-// this crate. Used for string enums because their `_Custom` variant can't be
-// truly private (only `#[doc(hidden)]`).
-#[doc(hidden)]
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct PrivOwnedStr(Box<str>);
-
-#[cfg(not(tarpaulin_include))]
-impl std::fmt::Debug for PrivOwnedStr {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.0.fmt(f)
     }
 }